[Impacted Area Based PR testing] Fix actually count calculation for test scripts. (#16487)
What is the motivation for this PR?
In PR #16437, we addressed the script timing issue by selecting the five most recent test plans. However, this caused an incorrect ActualCount, as it was summarizing the number of test cases instead of the test scripts. Since the number of test scripts matches the number of test plans, this PR updates the query to correctly count the test plans.

How did you do it?
Since the number of test scripts matches the number of test plans, this PR updates the query to correctly count the test plans.

Signed-off-by: yenlu <yenlu@cisco.com>