[Chassis] Reduce wait/timeout time for chassis (#16077)
Description of PR
Cisco chassis used to be unstable months ago, so updated the wait time of sshd ready to 600s, and service ready to 900s.
After the image became stable, we can reduce the time to:
sshd ready: 420s
service ready: 600s

Will keep monitoring and see whether we can reduce the wait time to align with t0/t1 in the future.

Approach
What is the motivation for this PR?
Cisco chassis used to be unstable months ago, so updated the wait time of sshd ready to 600s, and service ready to 900s.
After the image became stable, we can reduce the time to:
sshd ready: 420s
service ready: 600s

Will keep monitoring and see whether we can reduce the wait time to align with t0/t1 in the future.

How did you do it?
How did you verify/test it?
In the nightly test, the configuration works well.

Any platform specific information?
Supported testbed topology if it's a new test case?
Documentation

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>