feat: add wait_for_bgp option to reboot (#16542) (#16562)
Description of PR
The reboot process for chassis will need longer time to wait for the BGP to be established. For example, the acl/test_acl.py can be flaky if we are not waiting BGP long enough. Therefore, we are introducing wait_for_bgp option to the reboot() function.

Summary:
Fixes # (issue) Microsoft ADO 30862178

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 New Test case
 Skipped for non-supported platforms
 Test case improvement
Back port request
 202012
 202205
 202305
 202311
 202405
 202411
Approach
What is the motivation for this PR?
We found that tests like acl/test_acl.py can be flaky if we are not waiting for BGP long enough on chassis after reboot. Therefore, we want to mimic what we have in config_reload() to also introduce the wait_for_bgp option to the reboot() function.

How did you do it?
How did you verify/test it?
I ran the updated acl test code and can make sure it's working well.

co-authorized by: jianquanye@microsoft.com