[Impacted Area Based PR testing] Improve PR checker naming and script filtering consistency. (#16591)
What is the motivation for this PR?
To prevent error handling in the Azure pipeline YAML, we avoid using overlapping PR checker names by adding a suffix for the T0 and T1 PR checkers. However, we continue using the topology name to filter test scripts, which no longer matches the PR checker name. To resolve this issue and standardize the approach, we apply the suffix to all PR checkers and filter the scripts using the topology_{suffix} format.

How did you do it?
To resolve this issue and standardize the approach, we apply the suffix to all PR checkers and filter the scripts using the topology_{suffix} format.

How did you verify/test it?
Test by pipeline itself, to see if we can filter the scripts correctly. All PR checkers can get their corresponding scripts.