Upgrade docker SDK client version
Previously the upgrade of requests package caused a compatibility issue with the docker SDK client.
A workaround was introduced in the playbook to force install a lower version requests package.

After the  issue was fixed in the docker SDK client side, the workaround code can be removed.

This change removed the workaround code and upgraded the docker client to version 7.1.0.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>