[Impacted Area Based PR testing] Roll out T0 and T1 PR checkers.  (#16598)
What is the motivation for this PR?
In PRs #15666 and #16403, we partially rolled out the T0 and T1 PR checkers, considering resource utilization since these checkers require over 20 instances and needed to run in parallel with the legacy PR checkers. After a period of observation, we have confirmed the stability of the new system. In this PR, we complete the rollout of the remaining T0 and T1 PR checkers and officially deprecate the old PR checkers. At the same time, we have added all test scripts into PR testing, and we will gather scripts though pytest mark, so we don't need onboarding PR checkers anymore.

How did you do it?
In this PR, we complete the rollout of the remaining T0 and T1 PR checkers and officially deprecate the old PR checkers.

How did you verify/test it?
Test by pipeline itself, to see if we can successfully pass the PR checkers.

Signed-off-by: yenlu <yenlu@cisco.com>