Upgrade docker SDK client version (#16593)
What is the motivation for this PR?
Previously the upgrade of requests package caused a compatibility issue with the docker SDK client. A workaround was introduced in the playbook to force install a lower version requests package.

After the issue was fixed in the docker SDK client side, the workaround code can be removed.

How did you do it?
This change removed the workaround code and upgraded the docker client to version 7.1.0.

How did you verify/test it?
Tested add/remove topo using the testbed-cli.sh

Signed-off-by: yenlu <yenlu@cisco.com>