upload fake dutConfig and dutQosConfig parameters to support kvm PR test (#16580)
What is the motivation for this PR?
TO support PR test of QoS SAI, need to provide varous configuration to mock dutConfig and dutQosConfig

How did you do it?
collect dutConfig and dutQosConfig on a single box 8102, and upload as fake parameter for mock purpose

How did you verify/test it?

Signed-off-by: Xu Chen <xuchen3@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>