Fix gnmi/test_gnoi_killprocess.py checking disabled services (#16619)
gnmi/test_gnoi_killprocess.py is a newly added test in 202411 and it
aims to test functionality of a new API - KillProcess. It tests the API
by killing then re-enabling services, but 2 of the tested services in
the test file is disabled in upstream 202411 build config (restapi and
telemtry), resulting in a `service could not be found` failure.

https://github.com/sonic-net/sonic-buildimage/blob/202411/rules/config#L134
https://github.com/sonic-net/sonic-buildimage/blob/202411/rules/config#L150

Removing the `telemetry` and `restapi` case from the test file so it
does not fail from testing disabled services. Also skipping the
dhcp_relay case for t1 topos as t1 topos does not use the dhcp_relay service
as well.

This test file no longer fails on `Unit XXXXX.service could not be
found`.