[dualtor] Fine-tune the mux simulator control post timeout (#16651)
This is to fix #16495.
The root cause is that, on dualtor testbed, the post request to mux_simulator to toggle all mux ports can take up to 15+ seconds to finish; but the post timeout is 10s, so the post request to toggle all mux ports always timeout.

Here is the workflow for a HTTP requests to mux_simulator:

#
# how does a HTTP request to mux_simulator work?
#         +------------+       +------------+      +-------------+       +--------------+
# ------->|listen queue+------>|accept queue+----->|mux simulator+------>| mux simulator|
#         +------------+       +------------+      |  dispatcher |       |handler worker|
#                                                  +-------------+       +--------------+
The HTTP requests are handled by TCP listen/accept queue first; after the connection is established, mux_simulator will consume the request and dispatch the request to workers to handle the request.
Back to the toggle all mux ports post request, the client times out and disconnects, but the request has already been dispatched by the mux_simulator, its handler worker is still busy handling the request. If our test code send more timeout requests to mux_simulator, those inundated requests could overload the mux_simulator handler workers and a kind of denial-of-service occurs. Even worse, with more HTTP requests coming, the listen/accept queue will overflow.

This PR is to stop creating timeout requests to mux_simulator.

Signed-off-by: Longxiang Lyu <lolv@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>