[Chassis][voq] Remove supervisorAsicBase offset from test_fabric_reach.py test (#15828)
Approach
What is the motivation for this PR?
The original test used a base moduleID offset for Arista devices, however, this offset created mismatch in the moduleID and led the test to fail.

How did you do it?
Remove the base module ID offset variable: supervisorAsicBase

How did you verify/test it?
I ran the test successfully on a T2 testbed to verify the change.

Signed-off-by: yenlu <yenlu@cisco.com>