Added test for ecn counter increment verification for already congestion experienced  packets (#16356)
Description of PR
Summary:
Fixes # (issue)

Approach
What is the motivation for this PR?
Verify if ingress congestion experienced markets pkts (ECT bit b11) are causing the ecn counter increment on the DUT as expected on congestion

How did you do it?
Wrote snappi test to cause congestion on the port

How did you verify/test it?
Injected frames with ECT bit set and caused congestion on the egress port.

Verified that the DUT is counting the ECN marking and the counter is incrementing.

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>