fix: pfc timer accuracy timestamp regex (#16757)
Description of PR
Summary:
Fixes # (issue) 31202914

From #16446 we are changing to use regex to capture the timestamp for pfc timer accuracy. However, the regex does not cover the scenario where date digit is 1. For example: 2025 Jan  5 00:53:14.103188

Approach
What is the motivation for this PR?
Comparing between:

2025 Jan 26 17:36:31.27789
2025 Feb  2 15:17:24.055182
We can see that we might need up to 2 spaces after month and the number of day digits can vary between 1 to 2.

How did you do it?
Adjust the regex so it matches for both format

How did you verify/test it?
image

Signed-off-by: Austin Pham <austinpham@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>