[ip/test_mgmt_ipv6_only]: Refactor to properly cleanup on failure (#16442)
Description of PR
ipv6 mgmt only tests can fail in setup, after config has been modified, leading to the teardown to not be run, and the duts to be left in an unreachable state. Refactor the test to split backing up and restoring the config into an independent fixture, such that failures in setting the mgmt IPs won't nuke the testbed. Also refactor / cleanup some other aspects of the tests to enhance clarity.

Summary:
Fixes #16441

Approach
What is the motivation for this PR?
Fix failures in test setup leaving the testbed in a wedged state, due to removed ipv4 mgmt IPs

How did you do it?
Refactor the test to have config backup and restore in a separate fixture

Signed-off-by: Liam Kearney <liamkearney@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>