Fix invalid parameter test cases in wol
Fix the invalid_count, invalid_interval, and invalid_password test cases which check
the response of invalid parameters are to the wol command.  For invalid_password,
updated to correspond to the changed error message.  For the other two passed in an
interval/count since the existence of all required params is now validated before
checking the validity of the parameter value.

Signed-off-by: Veronica Cojocaru <vcojocaru@arista.com>