Increase Orch CPU utilization timeout before link flap (#16187)
* Increase Orch CPU utilization timeout before link flap

This change was made because in modular chassis with multi-asic LCs, the
link flap test might run on the uplink LC followed by the downlink LC.
In such a scenario, the downlink LC will still be hot (above 10%
utilization) before we flap its interfaces. Hence, the increase in
timeout.

We tested it with a timeout of 500 and it failed so we are increasing it
to 600 which has been passing on our local T2 testbeds.

* Pre-flap timeout is increased only for T2 devices

* Modified comment to explain the increased timeout in T2

Signed-off-by: yenlu <yenlu@cisco.com>