fix(tests): add backplane port only in srv6 tests (#16669)
What is the motivation for this PR?
Backplane port is necessary in our srv6 tests, but it will cause random failure in other tests. When the test packet dst IP matches the IP prefix advertised by the exabgp, the ptf backplane interface will receive the test packet from the neighbor VM. The reason is the routes are advertised by exabgp to VM through the ptf backplane interface. And methods like verify_packet_any_port() not only validate the packet is received by the expected ports, but also validate it's not received by the unexpected ports.

How did you do it?
Added configuration options in ptf initial to limite the configuration scenarios of backplane ports only in srv6 tests

How did you verify/test it?
we tested it via daily jenkins run

Signed-off-by: linsongnan <linsongnan.lsn@alibaba-inc.com>