Retry gnmi_get calls for potential server startup timing issue (#16910)
What is the motivation for this PR?
There are cases where gnmi_get is called when server is not fully ready after rotation which is a timing issue for a few seconds. If we retry the gnmi_get call, it will succeed. Add wait_until to retry the client calls for a period of 30 seconds.

How did you do it?
Add wait_until to retry for 30 seconds

How did you verify/test it?
Manual test/pipeline