Merged PR 9691: Change list of test cases to exclude to list of test cases to run
Change list of test cases to exclude to list of test cases to run

Instead of specifying the upgrade test cases to exclude, instead specify
the test cases that should be run. That makes it easier to understand
what's actually going to run.

Also take this change to add in the new options for the iteration counts, and to add a description for the SAD cases field.

Most active pipeline files have been updated.

ADO: 31330992

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

----
#### AI description  (iteration 1)
#### PR Classification
Configuration update

#### PR Summary
This pull request changes the approach from specifying test cases to exclude to specifying test cases to run in various pipeline configuration files.
- `.azure-pipelines/sonic-metadata/arista/tbtk5-t0-7260-7.yml`: Replaced `SKIP_REBOOT_CASES` with `RUN_REBOOT_CASE` and updated related parameters.
- `.azure-pipelines/sonic-metadata/arista/vms20-t0-7060.yml`: Replaced `SKIP_REBOOT_CASES` with `RUN_REBOOT_CASE` and updated related parameters.
- `.azure-pipelines/sonic-metadata/arista/vms24-t0-7260-2.yml`: Replaced `SKIP_REBOOT_CASES` with `RUN_REBOOT_CASE` and updated related parameters.
- `.azure-pipelines/sonic-metadata/metadata_upgrade_template.yml`: Replaced `SKIP_REBOOT_CASES` with `RUN_REBOOT_CASE` and updated related parameters.
- `.azure-pipelines/sonic-metadata/upgrade_sonic_metadata.yml`: Replaced `SKIP_REBOOT_CASES` with `RUN_REBOOT_CASE` and updated related parameters.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->