Retry from the first provided password if none of the passwords work (#16870)
* Retry from the first provided password if none of the passwords work

In the DeviceConnection class, if none of the passwords appear to work,
then try again from the first provided password. The reason for this is
if a a device initially needs some password specified in the alternate
passwords list, but then later needs to use some earlier-specified
password (because of some config change), then connection attempts will
fail until a new DeviceConnection object is instantiated.

Instead, work around that by trying the passwords in a loop (i.e. from
the beginning again). This also means that the class doesn't really need
to keep track of what password might be the "primary" password and what
password might be alternates.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Remove extra self

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

---------

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>