Added Enhancements for test_gnmi.py to run on different hwsku's on a testbed (#16955)
Description of PR
Enhanced test_gnmi.py to support execution across multiple HWSKUs on a testbed, rather than limiting it to a single HWSKU.

Previously, the test was designed to run on a randomly selected DUT using rand_one_dut_hostname, which could be either a line card or a supervisor node. However, an issue was observed where the test failed on the supervisor, and this was not detected until the test was executed as part of a nightly run that randomly selected a supervisor node.

To address this, the implementation has been updated to ensure that the test runs on both the supervisor and at least one of the line cards, improving coverage and reliability.

Summary:
Fixes # (issue)
ADO: 30909878

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 New Test case
 Skipped for non-supported platforms
 Test case improvement
Back port request
 202012
 202205
 202305
 202311
 202405
 202411
Approach
What is the motivation for this PR?
How did you do it?
Replacing rand_one_dut_hostname with enum_rand_one_per_hwsku_hostname

How did you verify/test it?
Ran on testbeds in Microsoft Lab

Signed-off-by: yenlu <yenlu@cisco.com>