Ensure minigraph restore when the test fails (#17019)
What is the motivation for this PR?
test_l2_configure moves the /etc/sonic/minigraph.xml, but when it fails, it won't restore the minigraph file.
let's ensure the minigraph is restored in this situation.

Signed-off-by: Longxiang Lyu lolv@microsoft.com

How did you do it?
Wrap the test operations in a try ... finally block, and put the minigraph restore in the finally clause.

How did you verify/test it?
Run on dualtor testbed.

Signed-off-by: Longxiang Lyu <lolv@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>