[202411] Remove onboarding test jobs and test scripts (#17044)
What is the motivation for this PR?
We have added lots of control plane and data plane test to PR test in master branch, and onboarding test job was created for piloting.
Currently we have finished the job in master branch, if we need to backport to 202411, will do cherry-pick and use impact area for PR test, do not need onboarding test job anymore.

How did you do it?
Remove onboarding test jobs and test scripts in 202411 branch