Replace testgap template with the 01-testcase.yml a unified template (#17079)
What is the motivation for this PR?
Remove the 02-testgap.yml template as the 01-testcase.yml is a unified template for bug, test gap and enhancement.

How did you do it?
Remove the 02-testgap.yml template as the 01-testcase.yml is a unified template for bug, test gap and enhancement.

How did you verify/test it?
NA

Signed-off-by: yenlu <yenlu@cisco.com>