Gnmi Test Case Fix (#17010)
Description of PR
Summary:
Fixes # (issue)
ADO: 30909878

Type of change
 Bug fix

Approach
What is the motivation for this PR?
The test will fail if the dut time is not in sync with the NTP Server because the gnmi certs have dependency upon system time.

How did you do it?
Syncing the dut time with NTP server while applying the certs

How did you verify/test it?
Ran it on Microsoft Lab testbed

---------

Co-authored-by: yatishkoul <yatishkoul@gmail.com>
Signed-off-by: yenlu <yenlu@cisco.com>