Gnmi Test Case Fix
<!--
Please make sure you've read and understood our contributing guidelines;
https://github.com/sonic-net/SONiC/blob/gh-pages/CONTRIBUTING.md

Please provide following information to help code review process a bit easier:
-->
### Description of PR
<!--
The test failed on the supervisor node because the DUT's system time was not synchronized with the NTP server. Since the GNMI certificates have a dependency on the system time, this discrepancy caused issues during the certificate application process. To resolve this, the DUT's time is synchronized with the NTP server before authentication part.
-->

Summary:
Fixes # (issue)
ADO: 30909878
### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [x] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
    - [ ] Skipped for non-supported platforms
- [ ] Test case improvement

### Back port request
- [] 202012
- [] 202205
- [] 202305
- [] 202311
- [x] 202405
- [x] 202411

### Approach
#### What is the motivation for this PR?
The test will fail if the dut time is not in sync with the NTP Server because the gnmi certs have dependency upon system time.
#### How did you do it?
Syncing the dut time with NTP server while applying the certs
#### How did you verify/test it?
Ran it on Microsoft Lab testbed
#### Any platform specific information?

#### Supported testbed topology if it's a new test case?

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->