[testbed_doc] Design doc for deploying a single testbed with multiple servers (#15395)
What is the motivation for this PR?
When deploying a testbed with a vast number of virtual ceos neighbors, we will create ceos containers on same server, however, the server doesn't have infinite resources such as memory to deploy that.
To leverage the servers instead of a single server, we proposal this design for deploying testbeds with multiple servers.

How did you do it?
Design for deploying testbed with multiple servers.

Signed-off-by: yenlu <yenlu@cisco.com>