Skip warm reboot related tests in 202412 testing (#17112)
What is the motivation for this PR?
Warmboot is not required for 202412 PR testing and currently fails during the process. To address this, we will bypass it in 202412 PR testing.

How did you do it?
Use conditional mark to skip test case platform_tests/test_advanced_reboot.py::test_warm_reboot.