Switch from tcpdump to dumpcap
This is (hopefully) more efficient on the system by being a single
process (but multiple threads for each interface), and also stores
information about which interface each packet was seen on. This also
eliminates the need to run mergecap afterwards.

In addition, specify the snapshot length and the buffer size to optimize
for small packets, but maybe a large burst of them at times.

For now, capture_filtered.pcap will still be a regular pcap file;
if/when tshark is added to the PTF container, tshark can then be used
for creating this file and still preserve interface information.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>