[202411][sonic-mgmt]Fix test_vxlan_hash in fib/test_fib.py for dualtor active-active topologies (#16949)
What is the motivation for this PR?
This test fails on DualTor

How did you do it?
The approach to fix this is to replace the ptf_test_port_map fixture with dualtor compatible fixture  ptf_test_port_map_active_active

How did you verify/test it?
Ran tests on Arista-7260CX3 in Dualtor Setup

Signed-off-by: yenlu <yenlu@cisco.com>