Use pcap for PTF sockets (which in turn uses mmap'ed sockets)
This is done by defining a new "remote-pcap" platform that PTF can use.
This is basically to describe how PTF should send and receive packets in
a particular environment. Right now, we've defined the remote platform
that uses Scapy with its defaults. This PR defines a new remote-pcap
platform where Scapy is configured to use pcap.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>