adding eos support (#17155)
chore: update test for t2

Description of PR
Summary:
Previously this test skipped for eos neighbor due to legacy. Now we should be able to enable this. This PR added the necessary changes to make this compatible

Fixes # (issue) 28839241

Approach
What is the motivation for this PR?
Described above

How did you do it?
Adjust the relevant code to be compatible with eos neighbor.

How did you verify/test it?
Verified on T2 testbed compatibility.

Signed-off-by: Austin Pham <austinpham@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>