Support baseline test in 202411 branch (#17154)
What is the motivation for this PR?
We are preparing to deploy impact area PR testing in 202411 branch, which won't always test all test sets, which means some test may not be test for a long time and if it's flaky, we are not able to find out. So baseline test is needed to find out flaky tests in 202411.

How did you do it?
Deploy baseline test in 202411