[XFAIL] xfail test_lag_member_forwarding.py for platform (#17305)
The feature is not ready for BRCM platform, need to xfail for BRCM before the fix is in place.

What is the motivation for this PR?
To xfail the test_lag_member_forwarding on BRCM platform.

How did you do it?
Xfail the test_lag_member_forwarding in conditional mark file.

How did you verify/test it?
Run test.

Any platform specific information?
Only xfail for BRCM platform.

Signed-off-by: yenlu <yenlu@cisco.com>