Merged PR 10225: Skip test_restapi as it's not required on T0
Skip test_restapi as it's not required on T0

This PR is to skip `test_restapi` as it's not required for T0 according discussion in ADO https://msazure.visualstudio.com/One/_workitems/edit/29844291

----
#### AI description  (iteration 1)
#### PR Classification
Test modification

#### PR Summary
This pull request updates the test configuration to skip the `test_restapi` as it is not required on T0 and needs improvement for T1.
- `tests/common/plugins/conditional_mark/tests_mark_conditions.yaml`: Updated the skip reason for `restapi/test_restapi.py` to indicate it is not required on T0 and needs improvement for T1.

Cherry picked from !9497

----
#### AI description  (iteration 1)
#### PR Classification
Test modification

#### PR Summary
This pull request updates the test configuration to skip the `test_restapi` as it is not required on T0 and needs improvement for T1.
- `tests/common/plugins/conditional_mark/tests_mark_conditions.yaml`: Updated the skip reason for `restapi/test_restapi.py` to indicate it is not required on T0 and needs improvement for T1.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->