Fix test_copp on non t2 topologies (#17314)
What is the motivation for this PR?
BGP is shutdown on a TOR different from the one where package download is supposed to happen in DualTOR topology. The cause of the issue is T2 approach is being extended to multi DUT topologies which is leading to failures.

How did you do it?
The DUT where package download is supposed to happen should match the one where BGP is being shutdown. In multi DUT topologies with this change, we are making sure that BGP shutdown happens in the appropriate DUT.

How did you verify/test it?
Test is consistently passing with the fix. Tested with 202411 image on Arista-7050CX3 with dualtor-aa.