Fix BMP tests (#17179)
What is the motivation for this PR?
bmp/test_bmp_statedb.py and bmp/test_docker_restart.py are failing as they try to test BMP without enabling it.

How did you do it?
Created a fixture to enable BMP. For the CLI to work we also need: sonic-net/sonic-buildimage#21863

How did you verify/test it?
Ran all tests in the BMP directory.

Signed-off-by: yenlu <yenlu@cisco.com>