[bugfix] Fix syntax error in conditional evaluation due to missing quotation mark (#17451)
What is the motivation for this PR?
The error

Traceback (most recent call last):
  File "/var/src/sonic-mgmt/tests/common/plugins/conditional_mark/__init__.py", line 549, in evaluate_condition
    condition_result = bool(eval(condition_str, basic_facts))
  File "<string>", line 1
    topo_name not in ['t0', 't0-64', 't0-116', 't0-118', 't0-35', 't0-56', 't0-standalone-32', 't0-standalone-64', 't0-standalone-128', 't0-standalone-256', 'dualtor-56', 'dualtor-120', 'dualtor', 't0-80', 't0-backend', 't1-lag', 't1-28-lag', t1-64-lag', 't1-56-lag', 't1-backend', 't2', 't2_2lc_36p-masic', 't2_2lc_min_ports-masic'] and asic_type not in ['mellanox']
is caused by the lack of '. In this PR, we fix this issue.

How did you do it?
Add a ' in the condition.

How did you verify/test it?
Test by pipeline, this error will not appear again.

Signed-off-by: yenlu <yenlu@cisco.com>