Merged PR 10397: Fixed a timing issue causing 'test_vxlan_priority_multi_pri_sec_switchover' to fail
After adding the IP address of VNET destination and one of its primary endpoints to `/tmp/vnet_monitor_block_ips.txt` on the PTF container, it takes some time for the `vnet_ping_responder.py` script to start blocking the IP address of the primary endpoint. The current delay in the code is 10 seconds, which is not enough since the test fails on some platforms (e.g., msn4600c) with the following message:
`RuntimeError: We received traffic with a downed endpoint(100.0.1.1), unexpected.`
This PR increases the wait time to 15 seconds. The test was run with the new changes on msn4600c and it passed.
Work items: [31744293](https://dev.azure.com/msazure/One/_workitems/edit/31744293) and [31220213](https://dev.azure.com/msazure/One/_workitems/edit/31220213)

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix for a failing test.

#### PR Summary
This pull request addresses a timing issue causing the `test_vxlan_priority_multi_pri_sec_switchover` to fail.
- `tests/vxlan/test_vxlan_ecmp_vnet_ping.py`: Added cleanup method for VNET routes in case of test failure.
- `tests/vxlan/test_vxlan_ecmp_vnet_ping.py`: Increased sleep duration from 10 to 15 seconds to ensure proper timing for test execution.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->