Merged PR 10039: Support Ad-hoc hops in multi-hop test (including 202405)
ADO: https://msazure.visualstudio.com/One/_workitems/edit/31240934

Took the opportunity as part of adding support for 202405 hop to rework how the upgrade path is specified. This will minimise the effort needed to add subsequent hops in the future (should be as simple if not simpler than A->B updates). There are now the following form fields:
![image.png](https://dev.azure.com/mssonic/12b9cbf4-b1d3-4768-8e49-669345c32e5d/_apis/git/repositories/5380e8f7-6e2a-4154-8dee-f3be7b096894/pullRequests/10039/attachments/image.png)

The base path can be specified as well as additional hops to run after the base path. This is a list where the hop locations and hop image numbers get paired (zipped) together to form the URL for the hop.

This replaces the old way of building the multi-hop path in sonic-metadata. Here's the PR to remove it: https://dev.azure.com/mssonic/internal/_git/sonic-metadata/pullrequest/10047

**Testing**
Multi-hop from 201811->202012->202305->202311->202405
https://dev.azure.com/mssonic/internal/_build/results?buildId=781850

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request adds support for ad-hoc hops in multi-hop upgrade tests.
- Added new script `assemble-multihop-upgrade-path.py` to assemble image URLs for multi-hop upgrade paths.
- Updated `.azure-pipelines/sonic-metadata/multi_hop_upgrade.yml` to include parameters for additional upgrade hop locations and image numbers.
- Modified `.azure-pipelines/sonic-metadata/upgrade_sonic_metadata.yml` to handle multi-hop upgrade paths and integrate the new script.
- Updated `.azure-pipelines/sonic-metadata/metadata_upgrade_template.yml` to support the new multi-hop upgrade test type.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->