Merged PR 10357: Update code for TK5 account access
To accommodate the additional privilege required for accessing the TK5 server, we created a separate account for TK5. In this PR, we update the code to align with this change.

#### Test
We can get the expected username and password
```
console_username = '_soniclab'
console_password ['<xpme password>']
```

----
#### AI description  (iteration 1)
#### PR Classification
Configuration update

#### PR Summary
This pull request updates the user name for console login in the configuration file.
- `ansible/group_vars/strtk5/secrets.yml`: Changed `console_login` from `console_login_tacacs` to `console_login_xpme`.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->