Modify wr_arp PTF script to generate an ARP packet of only 42 bytes (#17340)
In KVM environments, it's possible that an ARP response that gets
forwarded is less than 60 bytes, less than the minimum Ethernet frame
size. In physical environments, it'll almost always be at least the
minimum Ethernet frame size.

In the lab, the `arp_responder.py` script adds a marker in the padding
at the end of the ARP packet to indicate that it came from that script.
To gracefully handle those packets as well as actual ARP responses
generated by the device, generate the packet as a 42-byte packet, and
ignore extra bytes.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>