Fix the issue introduced by the PR #16716 (#17516)
Description of PR
Summary: After the #16716 merged, the remove topo failed. fix it in this PR
Fixes # (issue)

Approach
What is the motivation for this PR?
How did you do it?
How did you verify/test it?
Deploy the setup could pass with the fix

co-authorized by: jianquanye@microsoft.com

Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>