disable killprocess test to rewrite (#17507)
What is the motivation for this PR?
The GNOI KillProcess test was supposed to test whether GNOI can issue sigterm to a process, but since this is written to kill all sonic critical services and restarts, it ended up being very noisy with all issues on startup scripts. Disabling this test for now to rewrite and put the test in the scope of individual service instead of GNOI.

How did you do it?
Disable the test kill_process_and_restart

How did you verify/test it?
on KVM