fix: handle UnpicklingError in cache read (#17536)
Description of PR
Handle UnpicklingError edge case in FactsCache::read() when parallel run is enabled.

Summary:
Fixes # (issue) Microsoft ADO 31839137

Approach
What is the motivation for this PR?
When parallel run is enabled, multiple processes may try to read/write the same cache file, so there will be a chance that a file is being read by multiple processes at the same time, causing UnpicklingError in some of the processes. Therefore, we decided to retry to read the file after a short random sleep. If we still get the same error after retrying, we will return NOTEXIST to overwrite the file.

co-authorized by: jianquanye@microsoft.com